# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CalAI - Food Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface FoodAnalysisInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** MIME type of the image (e.g. image/jpeg) */
  mimeType: string;
  /**
     * Optional description of the food and quantity
     * @nullable
     */
  description?: string | null;
}

/**
 * Confidence level of the analysis
 */
export type FoodAnalysisResultConfidence = typeof FoodAnalysisResultConfidence[keyof typeof FoodAnalysisResultConfidence];


export const FoodAnalysisResultConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface FoodAnalysisResult {
  /** Identified food name */
  foodName: string;
  /** Total calories (kcal) */
  calories: number;
  /** Protein in grams */
  protein: number;
  /** Carbohydrates in grams */
  carbs: number;
  /** Fat in grams */
  fat: number;
  /** Fiber in grams */
  fiber: number;
  /** Estimated serving size description */
  servingSize: string;
  /** Confidence level of the analysis */
  confidence: FoodAnalysisResultConfidence;
  /** List of identified ingredients */
  ingredients: string[];
}

export type FoodLogInputMealType = typeof FoodLogInputMealType[keyof typeof FoodLogInputMealType];


export const FoodLogInputMealType = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner',
  snack: 'snack',
} as const;

export interface FoodLogInput {
  foodName: string;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  fiber: number;
  servingSize: string;
  mealType: FoodLogInputMealType;
  /**
     * Optional base64 thumbnail image
     * @nullable
     */
  imageBase64?: string | null;
  ingredients?: string[];
}

export type FoodLogMealType = typeof FoodLogMealType[keyof typeof FoodLogMealType];


export const FoodLogMealType = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner',
  snack: 'snack',
} as const;

export interface FoodLog {
  id: number;
  foodName: string;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  fiber: number;
  servingSize: string;
  mealType: FoodLogMealType;
  /** @nullable */
  imageBase64?: string | null;
  ingredients: string[];
  loggedAt: string;
}

export interface DailySummary {
  date: string;
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFat: number;
  totalFiber: number;
  entryCount: number;
}

export interface DayCalories {
  date: string;
  totalCalories: number;
  entryCount: number;
}

export type ListFoodLogsParams = {
/**
 * Filter by date (YYYY-MM-DD). Defaults to today if not provided.
 */
date?: string;
};