# lib/api-zod/src/generated/types/dayCalories.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CalAI - Food Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DayCalories {
  date: string;
  totalCalories: number;
  entryCount: number;
}