# lib/api-zod/src/generated/types/foodAnalysisInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CalAI - Food Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface FoodAnalysisInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** MIME type of the image (e.g. image/jpeg) */
  mimeType: string;
  /**
     * Optional description of the food and quantity
     * @nullable
     */
  description?: string | null;
}