# lib/api-zod/src/generated/types/foodAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CalAI - Food Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { FoodAnalysisResultConfidence } from './foodAnalysisResultConfidence';

export interface FoodAnalysisResult {
  /** Identified food name */
  foodName: string;
  /** Total calories (kcal) */
  calories: number;
  /** Protein in grams */
  protein: number;
  /** Carbohydrates in grams */
  carbs: number;
  /** Fat in grams */
  fat: number;
  /** Fiber in grams */
  fiber: number;
  /** Estimated serving size description */
  servingSize: string;
  /** Confidence level of the analysis */
  confidence: FoodAnalysisResultConfidence;
  /** List of identified ingredients */
  ingredients: string[];
}