# lib/api-zod/src/generated/types/foodLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CalAI - Food Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { FoodLogMealType } from './foodLogMealType';

export interface FoodLog {
  id: number;
  foodName: string;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  fiber: number;
  servingSize: string;
  mealType: FoodLogMealType;
  /** @nullable */
  imageBase64?: string | null;
  ingredients: string[];
  loggedAt: Date;
}