# lib/api-zod/src/generated/types/foodLogInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CalAI - Food Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { FoodLogInputMealType } from './foodLogInputMealType';

export interface FoodLogInput {
  foodName: string;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  fiber: number;
  servingSize: string;
  mealType: FoodLogInputMealType;
  /**
     * Optional base64 thumbnail image
     * @nullable
     */
  imageBase64?: string | null;
  ingredients?: string[];
}